feat: configure parser with option functions
Change parser configuration to use option functions instead of struct.
Also improve symbol filtering with filter functions which can be given
to a parser.

To make filter functions work, a new Symbol interface is introduced
which is implemented by the entities that can be filtered.

Also simplify CLI flags for filtering to work with new filter function
design.